requirements_dev.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
comm==0.1.3
db-dtypes==1.1.1
debugpy==1.6.7
decorator==5.1.1
deprecation==2.1.0
executing==1.2.0
Flask==2.3.2
frozenlist==1.4.0
fsspec==2023.6.0
functions-framework==3.4.0
gcsfs==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-cloud-appengine-logging==1.3.1
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.11.1
google-cloud-core==2.3.2
google-cloud-logging==3.6.0
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.2
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.25.0
oauthlib==3.2.2
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.6.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.23.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.16
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.15.0
